Splits up TKE budget calculation into multiple files and adds shear production
